fix build enigma2-plugin-extensions-multiquickbutton then po files is missing

+++ recipes-oe-alliance/e2openplugins/enigma2-plugin-extensions-multiquickbutton.bb
@@ -44,7 +44,9 @@ do_install_po() {
 	install -d ${D}${datadir}/enigma2/po/
 	LANGS="ar bg ca cs da de el en en_GB es et fa fi fr fy he hr hu is it lt lv nl no pl pt ru sv sk sr th tr uk"
 	for lang in ${LANGS}; do
-		install -m 0755 ${S}/po/$lang.po ${D}${datadir}/enigma2/po/multiquickbutton-$lang.po
+		if [ -e ${S}/po/$lang.po ]; then
+			install -m 0755 ${S}/po/$lang.po ${D}${datadir}/enigma2/po/multiquickbutton-$lang.po
+		fi	
 	done
 }
